requirements.txt
streamlit
google-analytics-data
google-auth
supabase
pandas
